// frontend/src/types/Penjualan.ts
export type Penjualan = {
  _id: string
  id_pos: string
  total_semua: string
  diskon: string
  bayar: string
  tanggal_mulai?: string
  tanggal_akhir?: string
  via: string
  piutang: string
  id_pelanggan?: string
  inv?: string
  selisih?: number
  id_harga?: string
  id_outlet?: string
  catatan?: string
  sub_total?: string
  nama?: string
}